+++ src/main/java/com/clickhouse/kafka/connect/sink/ClickHouseSinkConfig.java
@@ -1,16 +1,17 @@
 package com.clickhouse.kafka.connect.sink;
 
-import com.clickhouse.kafka.connect.sink.state.provider.KeeperStateProvider;
 import org.apache.kafka.common.config.ConfigDef;
-import org.apache.kafka.common.config.ConfigException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import java.io.UnsupportedEncodingException;
+import java.nio.charset.StandardCharsets;
+import java.util.HashMap;
 import java.util.Map;
 
 public class ClickHouseSinkConfig {
     private static final Logger LOGGER = LoggerFactory.getLogger(ClickHouseSinkConfig.class);
+
+    //Configuration Names
     public static final String HOSTNAME = "hostname";
     public static final String PORT = "port";
     public static final String DATABASE = "database";
@@ -20,7 +21,12 @@ public class ClickHouseSinkConfig {
     public static final String TIMEOUT_SECONDS = "timeoutSeconds";
     public static final String RETRY_COUNT = "retryCount";
     public static final String EXACTLY_ONCE = "exactlyOnce";
+    public static final String SUPPRESS_TABLE_EXISTENCE_EXCEPTION = "suppressTableExistenceException";
 
+
+
+
+    
     public static final int MILLI_IN_A_SEC = 1000;
     private static final String databaseDefault = "default";
     public static final int portDefault = 8443;
@@ -37,30 +43,26 @@ public class ClickHouseSinkConfig {
         KEEPER_MAP
     }
 
-    private Map<String, String> settings = null;
-    private String hostname;
-    private int port;
-    private String database;
-    private String username;
-    private String password;
-    private boolean sslEnabled;
-    private boolean exactlyOnce;
+    private final String hostname;
+    private final int port;
+    private final String database;
+    private final String username;
+    private final String password;
+    private final boolean sslEnabled;
+    private final boolean exactlyOnce;
+    private final int timeout;
+    private final int retry;
+    private final boolean suppressTableExistenceException;
 
-    private int timeout;
-
-    private int retry;
+    private final Map<String, String> clickhouseSettings;
 
     public static class UTF8String implements ConfigDef.Validator {
 
         @Override
         public void ensureValid(String name, Object o) {
             String s = (String) o;
-            try {
-                if (s != null ) {
-                    byte[] tmpBytes = s.getBytes("UTF-8");
-                }
-            } catch (UnsupportedEncodingException e) {
-                throw new ConfigException(name, o, "String must be non-empty");
+            if (s != null ) {
+                byte[] tmpBytes = s.getBytes(StandardCharsets.UTF_8);
             }
         }
 
@@ -81,8 +83,38 @@ public class ClickHouseSinkConfig {
         timeout = Integer.parseInt(props.getOrDefault(TIMEOUT_SECONDS, timeoutSecondsDefault.toString())) * MILLI_IN_A_SEC; // multiple in 1000 milli
         retry = Integer.parseInt(props.getOrDefault(RETRY_COUNT, retryCountDefault.toString()));
         exactlyOnce = Boolean.parseBoolean(props.getOrDefault(EXACTLY_ONCE,"false"));
-        LOGGER.info("exactlyOnce: " + exactlyOnce);
-        LOGGER.info("props: " + props);
+        suppressTableExistenceException = Boolean.parseBoolean(props.getOrDefault("suppressTableExistenceException","false"));
+
+        Map<String, String> clickhouseSettings = new HashMap<>();
+        String clickhouseSettingsString = props.getOrDefault("clickhouseSettings", "").trim();
+
+        if (!clickhouseSettingsString.isBlank()) {
+            String [] stringSplit = clickhouseSettingsString.split(",");
+            for (String clickProp: stringSplit) {
+                String [] propSplit = clickProp.trim().split("=");
+                if ( propSplit.length == 2 ) {
+                    clickhouseSettings.put(propSplit[0].trim(), propSplit[1].trim());
+                }
+            }
+        }
+        this.clickhouseSettings = clickhouseSettings;
+        this.addClickHouseSetting("insert_quorum", "2", false);
+        this.addClickHouseSetting("input_format_skip_unknown_fields", "1", false);
+        this.addClickHouseSetting("send_progress_in_http_headers", "1", false);
+
+        LOGGER.debug("ClickHouseSinkConfig: hostname: {}, port: {}, database: {}, username: {}, sslEnabled: {}, timeout: {}, retry: {}, exactlyOnce: {}",
+                hostname, port, database, username, sslEnabled, timeout, retry, exactlyOnce);
+        LOGGER.debug("ClickHouseSinkConfig: clickhouseSettings: {}", clickhouseSettings);
+    }
+
+    public void addClickHouseSetting(String key, String value, boolean override) {
+        if (clickhouseSettings.containsKey(key)) {
+            if (override) {
+                clickhouseSettings.put(key, value);
+            }
+        } else {
+            clickhouseSettings.put(key, value);
+        }
     }
 
     public static final ConfigDef CONFIG = createConfigDef();
@@ -178,6 +210,15 @@ public class ClickHouseSinkConfig {
                 ++orderInGroup,
                 ConfigDef.Width.MEDIUM,
                 "enable exactly once semantics.");
+        configDef.define(SUPPRESS_TABLE_EXISTENCE_EXCEPTION,
+                ConfigDef.Type.BOOLEAN,
+                false,
+                ConfigDef.Importance.LOW,
+                "suppress table existence exception. default: false",
+                group,
+                ++orderInGroup,
+                ConfigDef.Width.SHORT,
+                "suppress table existence exception.");
 
         return configDef;
     }
@@ -205,10 +246,14 @@ public class ClickHouseSinkConfig {
     public boolean isSslEnabled() {
         return sslEnabled;
     }
-
     public int getTimeout() {
         return timeout;
     }
     public int getRetry() { return retry; }
     public boolean getExactlyOnce() { return exactlyOnce; }
+    public boolean getSuppressTableExistenceException() {
+        return suppressTableExistenceException;
+    }
+    public Map<String, String> getClickhouseSettings() {return clickhouseSettings;}
+
 }