+++ src/test/java/com/clickhouse/kafka/connect/sink/ClickHouseSinkTaskWithSchemaTest.java
@@ -7,6 +7,8 @@ import org.apache.kafka.common.record.TimestampType;
 import org.apache.kafka.connect.data.Schema;
 import org.apache.kafka.connect.data.SchemaBuilder;
 import org.apache.kafka.connect.data.Struct;
+import org.apache.kafka.connect.data.Time;
+import org.apache.kafka.connect.data.Timestamp;
 import org.apache.kafka.connect.errors.DataException;
 import org.apache.kafka.connect.sink.SinkRecord;
 import org.junit.jupiter.api.AfterAll;
@@ -30,7 +32,7 @@ public class ClickHouseSinkTaskWithSchemaTest {
     private static ClickHouseHelperClient chc = null;
 
     @BeforeAll
-    private static void setup() {
+    public static void setup() {
         db = new ClickHouseContainer("clickhouse/clickhouse-server:22.5");
         db.start();
 
@@ -390,6 +392,12 @@ public class ClickHouseSinkTaskWithSchemaTest {
                 .field("date32_number", Schema.OPTIONAL_INT32_SCHEMA)
                 .field("datetime_number", Schema.INT64_SCHEMA)
                 .field("datetime64_number", Schema.INT64_SCHEMA)
+                .field("timestamp_int64",  Timestamp.SCHEMA)
+                .field("timestamp_date", Timestamp.SCHEMA)
+                .field("time_int32" , Time.SCHEMA)
+                .field("time_date32" , Time.SCHEMA)
+                .field("date_date", Time.SCHEMA)
+                .field("datetime_date", Timestamp.SCHEMA)
                 .build();
 
 
@@ -411,6 +419,12 @@ public class ClickHouseSinkTaskWithSchemaTest {
                     .put("date32_number", localDateInt)
                     .put("datetime_number", localDateTimeLong)
                     .put("datetime64_number", currentTime)
+                    .put("timestamp_int64", new Date(System.currentTimeMillis()))
+                    .put("timestamp_date",  new Date(System.currentTimeMillis()))
+                    .put("time_int32", new Date(System.currentTimeMillis()))
+                    .put("time_date32", new Date(System.currentTimeMillis()))
+                    .put("date_date", new Date(System.currentTimeMillis()))
+                    .put("datetime_date", new Date(System.currentTimeMillis()))
                     ;
 
 
@@ -624,7 +638,7 @@ public class ClickHouseSinkTaskWithSchemaTest {
 
         String topic = "support-dates-table-test";
         dropTable(chc, topic);
-        createTable(chc, topic, "CREATE TABLE `%s` ( `off16` Int16, date_number Nullable(Date), date32_number Nullable(Date32), datetime_number DateTime, datetime64_number DateTime64 ) Engine = MergeTree ORDER BY off16");
+        createTable(chc, topic, "CREATE TABLE `%s` ( `off16` Int16, date_number Nullable(Date), date32_number Nullable(Date32), datetime_number DateTime, datetime64_number DateTime64, timestamp_int64 Int64, timestamp_date DateTime64, time_int32 Int32, time_date32 Date32, date_date Date, datetime_date DateTime ) Engine = MergeTree ORDER BY off16");
         // https://github.com/apache/kafka/blob/trunk/connect/api/src/test/java/org/apache/kafka/connect/data/StructTest.java#L95-L98
         Collection<SinkRecord> sr = createDateType(topic, 1);
 
@@ -649,7 +663,7 @@ public class ClickHouseSinkTaskWithSchemaTest {
 
         String topic = "support-unsupported-dates-table-test";
         dropTable(chc, topic);
-        createTable(chc, topic, "CREATE TABLE `%s` ( `off16` Int16, date_number Date, date32_number Date32, datetime_number DateTime, datetime64_number DateTime64 ) Engine = MergeTree ORDER BY off16");
+        createTable(chc, topic, "CREATE TABLE `%s` ( `off16` Int16, date_number Date, date32_number Date32, datetime_number DateTime, datetime64_number DateTime64) Engine = MergeTree ORDER BY off16");
 
         Collection<SinkRecord> sr = createUnsupportedDataConversions(topic, 1);