+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+## 0.0.19 2023-08-23
+* Additional tests for ExactlyOnce
+* Allows customized settings (WIP)
+* Tweaked deduplication behavior to account for dynamic fields
+
+## 0.0.18 2023-07-17
+* Support inline schema with org.apache.kafka.connect.data.Timestamp type 
+* Support inline schema with org.apache.kafka.connect.data.Time type
+
+## 0.0.17 2023-07-12
+* Updating Logo
+
 ## 0.0.16 2023-06-13
 * Updated state handling so that warnings are posted, rather than errors + exceptions.
 * Adding 285 TOO_FEW_LIVE_REPLICAS to the retry list